% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNullModel.R
\name{fitNullModel}
\alias{fitNullModel}
\title{Fit null model}
\usage{
fitNullModel(y, X, covMatList = NULL, group.idx = NULL,
  family = "gaussian", AIREML.tol = 1e-06, maxIter = 100,
  dropZeros = TRUE, verbose = TRUE)
}
\arguments{
\item{y}{outcome vector}

\item{X}{data.frame or model.matrix}

\item{covMatList}{A list of matrices specifying the covariance structures of the random effects terms}

\item{group.idx}{group variable (index?)}

\item{family}{A description of the error distribution to be used in the model. The default "gaussian" fits a linear mixed model; see \code{\link{family}} for further options.}

\item{AIREML.tol}{The convergence threshold for the Average Information REML (AIREML) procedure used to estimate the variance components of the random effects.}

\item{maxIter}{The maximum number of iterations allowed in the AIREML procedure}

\item{dropZeros}{Logical indicator of whether variance component terms that converge to 0 should be removed from the model}

\item{verbose}{Logical indicator of whether updates from the function should be printed to the console}
}
\value{
The null model (TODO fill in details)
}
\description{
Fit null model
}
